refactor: move api and use queryKey factor
