+++ src/library-authoring/LibraryHome.tsx
@@ -7,7 +7,7 @@ import {
 import { NoComponents, NoSearchResults } from './EmptyStates';
 import LibraryCollections from './LibraryCollections';
 import LibraryComponents from './LibraryComponents';
-import { useLibraryComponentCount } from './data/apiHook';
+import { useLibraryComponentCount } from './data/apiHooks';
 import messages from './messages';
 
 const Section = ({ title, children } : { title: string, children: React.ReactNode }) => (
