+++ src/library-authoring/create-library/data/apiHooks.ts
@@ -4,6 +4,7 @@ import {
 } from '@tanstack/react-query';
 
 import { createLibraryV2 } from './api';
+import { libraryAuthoringQueryKeys } from '../../data/apiHooks';
 
 /**
  * Hook that provides a "mutation" that can be used to create a new content library.
@@ -15,7 +16,7 @@ export const useCreateLibraryV2 = () => {
   return useMutation({
     mutationFn: createLibraryV2,
     onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['listV2Libraries'] });
+      queryClient.invalidateQueries({ queryKey: libraryAuthoringQueryKeys.contentLibraryList() });
     },
   });
 };
