+++ src/studio-home/__mocks__/index.js
@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as studioHomeMock } from './studioHomeMock';
-export { default as listStudioHomeV2LibrariesMock } from './listStudioHomeV2LibrariesMock';
