+++ src/studio-home/data/api.test.js
@@ -14,12 +14,10 @@ import {
   getStudioHomeCoursesV2,
   getStudioHomeLibraries,
 } from './api';
-import { getStudioHomeLibrariesV2 } from '../../library/data/api';
 import {
   generateGetStudioCoursesApiResponse,
   generateGetStudioHomeDataApiResponse,
   generateGetStudioHomeLibrariesApiResponse,
-  generateGetStudioHomeLibrariesV2ApiResponse,
 } from '../factories/mockApiResponses';
 
 let axiosMock;
@@ -75,16 +73,6 @@ describe('studio-home api calls', () => {
     axiosMock.onGet(apiLink).reply(200, generateGetStudioHomeLibrariesApiResponse());
     const result = await getStudioHomeLibraries();
     const expected = generateGetStudioHomeLibrariesApiResponse();
-
-    expect(axiosMock.history.get[0].url).toEqual(apiLink);
-    expect(result).toEqual(expected);
-  });
-
-  it('should get studio v2 libraries data', async () => {
-    const apiLink = `${getApiBaseUrl()}/api/libraries/v2/`;
-    axiosMock.onGet(apiLink).reply(200, generateGetStudioHomeLibrariesV2ApiResponse());
-    const result = await getStudioHomeLibrariesV2({});
-    const expected = generateGetStudioHomeLibrariesV2ApiResponse();
 
     expect(axiosMock.history.get[0].url).toEqual(apiLink);
     expect(result).toEqual(expected);
